Changes Merged:
	Made cmd variable a byte for binary operations
	SPY-51: Bug in OperationImpl's decodeLong(2)

+++ src/main/java/net/spy/memcached/protocol/binary/GetOperationImpl.java
@@ -30,7 +30,7 @@ import net.spy.memcached.ops.GetOperation;
  */
 class GetOperationImpl extends SingleKeyOperationImpl implements GetOperation {
 
-  static final int GET_CMD = 0x00;
+  static final byte GET_CMD = 0x00;
 
   /**
    * Length of the extra header stuff for a GET response.
